build apigateway by spring cloud gateway and config export api public and private
