+++ indentity-service/src/main/java/com/example/identity/dto/response/UserResponse.java
@@ -1,10 +1,9 @@
 package com.example.identity.dto.response;
 
-import java.time.LocalDate;
-import java.util.Set;
-
 import lombok.*;
 import lombok.experimental.FieldDefaults;
+
+import java.util.Set;
 
 @Data
 @Builder
@@ -16,8 +15,5 @@ import lombok.experimental.FieldDefaults;
 public class UserResponse {
     String id;
     String username;
-    String firstName;
-    String lastName;
-    LocalDate dob;
     Set<RoleResponse> roles;
 }
